+++ changelog.md
@@ -5,6 +5,10 @@
 
 # ChangeLog:
 
+#### v11.02
+
+- Fixed bugs
+
 #### v11.01
 
 - Refactored docker build script to limit build cache to 35gb